Move `toReadable` from `ModHashMap` to `RoomMods`
